Acquisition Function Base Class

Based on a surrogate model, the acquisition function encodes the preference to evaluate a new point for evaluation.

See also

Super class

bbotk::Objective -> AcqFunction

Public fields

direction

(character(1))
Optimization direction of the acquisition function relative to the direction of the objective function of the bbotk::OptimInstance. Must be "same", "minimize", or "maximize".

surrogate_max_to_min

(-1 | 1). Multiplicative factor to correct for minimization or maximization of the acquisition function.

Active bindings

archive

(bbotk::Archive)
Points to the bbotk::Archive of the Surrogate.

fun

(function)
Objective function.

surrogate

Surrogate.

Methods

Public methods

AcqFunction$new()

AcqFunction$update()

AcqFunction$eval_many()

AcqFunction$eval_dt()

AcqFunction$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$format()

bbotk::Objective$print()

Method new()

Creates a new instance of this R6 class.

Note that the surrogate can be initialized lazy and can later be set via the active binding.

Usage

AcqFunction$new(id, constants = ps(), surrogate, direction)

Arguments

id

(character(1)).

constants

(paradox::ParamSet).

surrogate

(NULL | Surrogate).

direction

(character(1)).

Method update()

Update the acquisition function.

Usage

AcqFunction$update()

Method eval_many()

Evaluates multiple input values received as a list, converted to a data.table() on the objective function. Missing columns in xss are filled with NAs in xdt.

Usage

AcqFunction$eval_many(xss)

Arguments

xss

(list())
A list of lists that contains multiple x values, e.g. list(list(x1 = 1, x2 = 2), list(x1 = 3, x2 = 4)).

Returns

data.table::data.table() that contains one y-column for single-criteria functions and multiple y-columns for multi-criteria functions, e.g. data.table(y = 1:2) or data.table(y1 = 1:2, y2 = 3:4).

Method eval_dt()

Evaluates multiple input values on the objective function supplied by the user.

Usage

AcqFunction$eval_dt(xdt)

Arguments

xdt

(data.table::data.table())
Set of untransformed points / points from the search space. One point per row, e.g. data.table(x1 = c(1, 3), x2 = c(2, 4)). Column names have to match ids of the search_space. However, xdt can contain additional columns.

Returns

data.table::data.table()] that contains one y-column for single-criteria functions and multiple y-columns for multi-criteria functions, e.g. data.table(y = 1:2) or data.table(y1 = 1:2, y2 = 3:4).

Method clone()

The objects of this class are cloneable with this method.

Usage

AcqFunction$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.