Acquisition Function Expected Improvement Per Second

Expected improvement per second.

It is assumed that calculations are performed on an bbotk::OptimInstanceSingleCrit. Additionally to target values of the codomain that should be minimized or maximized, the bbotk::Objective of the bbotk::OptimInstanceSingleCrit should return time values. The id of the column containing the time values must be passed as the time_id argument during construction.

References

Snoek, Jasper, Larochelle, Hugo, Adams, P R (2012). “Practical Bayesian Optimization of Machine Learning Algorithms.” In Pereira F, Burges CJC, Bottou L, Weinberger KQ (eds.), Advances in Neural Information Processing Systems, volume 25, 2951--2959.

See also

Super classes

bbotk::Objective -> bbotk::ObjectiveRFunDt -> mlr3mbo::AcqFunction -> AcqFunctionEIPS

Public fields

y_best

(numeric(1)).

surrogate_time

SurrogateSingleCrit.

Methods

Public methods

AcqFunctionEIPS$new()

AcqFunctionEIPS$setup()

AcqFunctionEIPS$update()

AcqFunctionEIPS$update_surrogate()

AcqFunctionEIPS$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$format()

bbotk::Objective$print()

bbotk::ObjectiveRFunDt$eval_dt()

bbotk::ObjectiveRFunDt$eval_many()

Method new()

Creates a new instance of this R6 class.

Usage

AcqFunctionEIPS$new(surrogate, surrogate_time, time_id)

Arguments

surrogate

(SurrogateSingleCrit).

surrogate_time

(SurrogateSingleCrit).

time_id

(character(1)).

Method setup()

Sets up the acquisition function.

Usage

AcqFunctionEIPS$setup(archive)

Arguments

archive

(bbotk::Archive).

Method update()

Updates acquisition function and sets y_best.

Usage

AcqFunctionEIPS$update(archive)

Arguments

archive

(bbotk::Archive).

Method update_surrogate()

Update the Surrogate model(s) with new data. FIXME: probably moves this simply to $update()

Usage

AcqFunctionEIPS$update_surrogate(archive)

Arguments

archive

(bbotk::Archive).

Method clone()

The objects of this class are cloneable with this method.

Usage

AcqFunctionEIPS$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

library(bbotk)

#> Loading required package: paradox

library(paradox)
library(data.table)
library(mlr3)
library(mlr3learners)

objective = ObjectiveRFunDt$new(
  fun = function(xdt) data.table(y = xdt$x ^ 2, time = abs(xdt$x)),
  domain = ps(x = p_dbl(lower = -10, upper = 10)),
  codomain = ps(y = p_dbl(tags = "minimize"))
)

instance = OptimInstanceSingleCrit$new(objective, terminator = trm("evals", n_evals = 10L))

learner = lrn("regr.km", covtype = "matern3_2", optim.method = "gen")
surrogate = SurrogateSingleCritLearner$new(learner)
surrogate_time = surrogate$clone(deep = TRUE)

acq_function = AcqFunctionEIPS$new(surrogate, surrogate_time = surrogate_time, time_id = "time")

bayesopt_soo(instance, acq_function)

#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  28.97921 
#>   - best initial criterion value(s) :  -18.93939 
#> 
#> 
#> Tue Oct  5 20:16:01 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    2.897921e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-1.893939e+01
#>       1 	-1.893646e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -1.893646e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	4.723672e+00	G[ 1] :	6.520428e-10
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:01 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  28.97921 
#>   - best initial criterion value(s) :  -10.04367 
#> 
#> 
#> Tue Oct  5 20:16:01 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    2.897921e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-1.004367e+01
#>       1 	-9.853236e+00
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -9.853236e+00
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	3.762060e+00	G[ 1] :	-2.477536e-08
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:01 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  37.72107 
#>   - best initial criterion value(s) :  -24.8903 
#> 
#> 
#> Tue Oct  5 20:16:01 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.772107e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-2.489030e+01
#>       1 	-2.484642e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -2.484642e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	5.811499e+00	G[ 1] :	-1.872584e-08
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:01 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  37.72107 
#>   - best initial criterion value(s) :  -12.52265 
#> 
#> 
#> Tue Oct  5 20:16:01 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.772107e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-1.252265e+01
#>       1 	-1.251951e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -1.251951e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	6.657409e+00	G[ 1] :	2.518468e-10
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:01 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  37.72107 
#>   - best initial criterion value(s) :  -29.00237 
#> 
#> 
#> Tue Oct  5 20:16:01 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.772107e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-2.900237e+01
#>       1 	-2.891285e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -2.891285e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	1.086180e+01	G[ 1] :	-3.240039e-09
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:01 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  37.72107 
#>   - best initial criterion value(s) :  -14.30106 
#> 
#> 
#> Tue Oct  5 20:16:02 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.772107e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-1.380932e+01
#>       1 	-1.380255e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -1.380255e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	9.778123e+00	G[ 1] :	7.123999e-09
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:02 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  39.95678 
#>   - best initial criterion value(s) :  -32.4246 
#> 
#> 
#> Tue Oct  5 20:16:02 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.995678e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-3.242460e+01
#>       1 	-3.241960e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -3.241960e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	1.805137e+01	G[ 1] :	1.201512e-08
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:02 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  39.95678 
#>   - best initial criterion value(s) :  -14.45286 
#> 
#> 
#> Tue Oct  5 20:16:02 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.995678e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-1.445030e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -1.445029e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	1.297440e+01	G[ 1] :	2.919130e-09
#> 
#> Solution Found Generation 1
#> Number of Generations Run 3
#> 
#> Tue Oct  5 20:16:02 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  39.95678 
#>   - best initial criterion value(s) :  -35.3265 
#> 
#> 
#> Tue Oct  5 20:16:02 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.995678e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-3.526374e+01
#>       1 	-3.526007e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -3.526007e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	2.497282e+01	G[ 1] :	8.297042e-10
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:02 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  39.95678 
#>   - best initial criterion value(s) :  -14.83631 
#> 
#> 
#> Tue Oct  5 20:16:02 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.995678e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-1.481287e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -1.481287e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	1.560921e+01	G[ 1] :	1.045303e-09
#> 
#> Solution Found Generation 1
#> Number of Generations Run 3
#> 
#> Tue Oct  5 20:16:02 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  39.95678 
#>   - best initial criterion value(s) :  -37.09206 
#> 
#> 
#> Tue Oct  5 20:16:02 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.995678e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-3.701607e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -3.701553e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	3.178463e+01	G[ 1] :	-6.826920e-09
#> 
#> Solution Found Generation 1
#> Number of Generations Run 3
#> 
#> Tue Oct  5 20:16:02 2021
#> Total run time : 0 hours 0 minutes and 0 seconds
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : gen 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern3_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 
#>   - parameters upper bounds :  39.95678 
#>   - best initial criterion value(s) :  -19.49605 
#> 
#> 
#> Tue Oct  5 20:16:02 2021
#> Domains:
#>  1.000000e-10   <=  X1   <=    3.995678e+01 
#> 
#> Data Type: Floating Point
#> Operators (code number, name, population) 
#> 	(1) Cloning........................... 	3
#> 	(2) Uniform Mutation.................. 	0
#> 	(3) Boundary Mutation................. 	0
#> 	(4) Non-Uniform Mutation.............. 	0
#> 	(5) Polytope Crossover................ 	0
#> 	(6) Simple Crossover.................. 	0
#> 	(7) Whole Non-Uniform Mutation........ 	0
#> 	(8) Heuristic Crossover............... 	0
#> 	(9) Local-Minimum Crossover........... 	0
#> 
#> HARD Maximum Number of Generations: 5
#> Maximum Nonchanging Generations: 2
#> Population size       : 4
#> Convergence Tolerance: 1.000000e-03
#> 
#> Using the BFGS Derivative Based Optimizer on the Best Individual Each Generation.
#> Not Checking Gradients before Stopping.
#> Not Using Out of Bounds Individuals and Not Allowing Trespassing.
#> 
#> Maximization Problem.
#> 
#> 
#> Generation#	    Solution Value
#> 
#>       0 	-1.805194e+01
#>       1 	-1.803734e+01
#> 
#> 'wait.generations' limit reached.
#> No significant improvement in 2 generations.
#> 
#> Solution Fitness Value: -1.803734e+01
#> 
#> Parameters at the Solution (parameter, gradient):
#> 
#>  X[ 1] :	6.580972e+00	G[ 1] :	1.038087e-08
#> 
#> Solution Found Generation 1
#> Number of Generations Run 4
#> 
#> Tue Oct  5 20:16:02 2021
#> Total run time : 0 hours 0 minutes and 0 seconds

#> <Archive>
#>           x       y   time           timestamp batch_nr acq_eips
#>  1:   1.449 2.1e+00  1.449 2021-10-05 20:16:01        1       NA
#>  2:  -1.336 1.8e+00  1.336 2021-10-05 20:16:01        1       NA
#>  3:   8.863 7.9e+01  8.863 2021-10-05 20:16:01        1       NA
#>  4:  -5.627 3.2e+01  5.627 2021-10-05 20:16:01        1       NA
#>  5:  -9.998 1.0e+02  9.998 2021-10-05 20:16:01        2     24.2
#>  6:   5.631 3.2e+01  5.631 2021-10-05 20:16:01        3     21.3
#>  7:   9.980 1.0e+02  9.980 2021-10-05 20:16:02        4      7.7
#>  8:  -7.950 6.3e+01  7.950 2021-10-05 20:16:02        5      6.0
#>  9:   0.041 1.7e-03  0.041 2021-10-05 20:16:02        6      4.2
#> 10:  -3.546 1.3e+01  3.546 2021-10-05 20:16:02        7      3.4