Black-Box objective function

Specifies a black-box function and archive for optimizers to act upon. It allows the basic operations of querying the objective at design points ($eval_batch()), storing the evaluations in an internal archive and querying the archive ($archive()).

Evaluations of points are performed in batches. Before and after a batch is evaluated, the Terminator is queried for the remaining budget. If the available budget is exhausted, an exception is raised, and no further evaluations can be performed from this point on.

The optimizer is also supposed to store its final result, consisting of a list of optimal values and stimated performance values, by calling the method instance$assign_result.

Format

R6::R6Class object.

Construction

inst = Objective$new param_set, terminator)

This defines the resampled performance of a learner on a task, a feasibility region for the parameters the tuner is supposed to optimize, and a termination criterion.

param_set :: paradox::ParamSet.

terminator :: Terminator.

Fields

param_set :: paradox::ParamSet; from construction.

terminator :: Terminator; from construction.

n_evals :: integer(1)
Number of configuration evaluations stored in the container.

start_time :: POSIXct(1)
Time the tuning was started. This is set in the beginning of $tune() of mlr3tuning::Tuner.

result :: named list()
Result of the tuning, i.e., the optimal configuration and its estimated performance:

"perf": Named vector of estimated performance values of the best configuration found.

"tune_x": Named list of optimal hyperparameter settings, without potential trafo function applied.

"params": Named list of optimal hyperparameter settings, similar to tune_x, but with potential trafo function applied. Also, if the learner had some extra parameters statically set before tuning, these are included here.

Methods

eval_batch(dt)
data.table::data.table() -> named list()
Evaluates all hyperparameter configurations in dt through resampling, where each configuration is a row, and columns are scalar parameters. Updates the internal mlr3::BenchmarkResult $bmr by reference, and returns a named list with the following elements:

"batch_nr": Number of the new batch. This number is calculated in an auto-increment fashion and also stored inside the mlr3::BenchmarkResult as column batch_nr

"perf": A data.table::data.table() of evaluated performances for each row of the dt. Has the same number of rows as dt, and the same number of columns as length of measures. Columns are named with measure-IDs. A cell entry is the (aggregated) performance of that configuration for that measure.

Before and after each batch-evaluation, the Terminator is checked, and if it is positive, an exception of class terminated_error is raised. This function should be internally called by the tuner.

best(measure = NULL)
(mlr3::Measure, character(1)) -> mlr3::ResampleResult
Queries the mlr3::BenchmarkResult for the best mlr3::ResampleResult according to measure (default is the first measure in $measures). In case of ties, one of the tied values is selected randomly.

archive(unnest = "no")
character(1) -> data.table::data.table()
Returns a table of contained resample results, similar to the one returned by mlr3::benchmark()'s $aggregate() method. Some interesting columns of this table are:

All evaluated measures are included as numeric columns, named with their measure ID.

tune_x: A list column that contains the parameter settings the tuner evaluated, without potential trafo applied.

params: A list column that contains the parameter settings that were actually used in the learner. Similar to tune_x, but with potential trafo applied. Also, if the learner had some extra parameters statically set before tuning, these are included here. unnest can have the values "no", "tune_x" or "params". If it is not set to "no", settings of the respective list-column are stored in separate columns instead of the list-column, and dependent, inactive parameters are encoded with NA.

assign_result(tune_x, perf)
(list, numeric) -> NULL
The tuner writes the best found list of settings and estimated performance values here. For internal use.

tune_x: Must be a named list of settings only of parameters from param_set and be feasible, untransformed.

perf : Must be a named numeric vector of performance measures, named with performance IDs, regarding all elements in measures.

Methods

Public methods

Objective$new()

Objective$format()

Objective$print()

Objective$eval_batch()

Objective$assign_result()

Objective$clone()

Method new()

Usage

Objective$new(fun, param_set, terminator, minimize = TRUE)

Method format()

Usage

Objective$format()

Method print()

Usage

Objective$print()

Method eval_batch()

Usage

Objective$eval_batch(dt)

Method assign_result()

Usage

Objective$assign_result(tune_x, perf)

Method clone()

The objects of this class are cloneable with this method.

Usage

Objective$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.