Abstract Terminator Class

Abstract Terminator class that implements the main functionality each terminator must have. A terminator is an object that determines when to stop the tuning.

Termination of tuning works as follows:

Evaluations in a tuner are performed in batches.

Before and after a batch evaluation the Terminator is checked, if it is positive, we stop.

The tuning algorithm itself might decide not to produce any more points, or even might decide to do a smaller batch in its last evaluation.

Therefore the following note seems in order: While it is definitely possible to execute a fine-grained control for termination, and for many tuners we can specify exactly when to stop, it might happen that a few too many or even a few too few evaluations are performed, especially if the batch size is larger than 1. So better check the size of the returned archive.

Format

R6::R6Class object.

Construction

t = Terminator$new(param_set = ParamSet$new())

param_set :: paradox::ParamSet
Set of control parameters for terminator.

Fields

param_set :: paradox::ParamSet

Methods

is_terminated(inst)
mlr3tuning::TuningInstance -> logical(1)
Is TRUE iff the termination criterion is positive. Must be implemented in a subclass.

See also

Methods

Public methods

Terminator$new()

Terminator$format()

Terminator$print()

Terminator$is_terminated()

Terminator$clone()

Method new()

Usage

Terminator$new(param_set = ParamSet$new())

Method format()

Usage

Terminator$format()

Method print()

Usage

Terminator$print()

Method is_terminated()

Usage

Terminator$is_terminated(inst)

Method clone()

The objects of this class are cloneable with this method.

Usage

Terminator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.