Terminator that stops after a number of evaluations

Class to terminate the tuning depending on the number of evaluations. An evaluation is defined by one resampling of a parameter value.

Format

R6::R6Class object inheriting from Terminator.

Construction

TerminatorEvals$new()
term("evals")

Parameters

n_evals :: integer(1)
Number of allowed evaluations, default is 100L

See also

Super class

mlr3mbo::Terminator -> TerminatorEvals

Methods

Public methods

TerminatorEvals$new()

TerminatorEvals$is_terminated()

TerminatorEvals$clone()

Inherited methods

mlr3mbo::Terminator$format()

mlr3mbo::Terminator$print()

Method new()

Usage

TerminatorEvals$new()

Method is_terminated()

Usage

TerminatorEvals$is_terminated(obj)

Method clone()

The objects of this class are cloneable with this method.

Usage

TerminatorEvals$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
library(mlr3tuning)
TerminatorEvals$new()
term("evals", n_evals = 5)
}