Syntactic Sugar Acquisition Function Construction

This function complements mlr_acqfunctions with functions in the spirit of mlr_sugar from mlr3.

Usage

acqf(.key, ...)

Arguments

.key

(character(1))
Key passed to the respective dictionary to retrieve the object.

...

(named list())
Named arguments passed to the constructor, to be set as parameters in the paradox::ParamSet, or to be set as public field. See mlr3misc::dictionary_sugar_get() for more details.

Value

AcqFunction

Examples

acqf("ei")
#> <AcqFunctionEI:acq_ei>
#> Domain:
#> <ParamSet(0)>
#> Empty.
#> Codomain:
#> <Codomain(0)>
#> Empty.
#> Constants:
#> <ParamSet(1)>
#>         id    class lower upper nlevels default  value
#>     <char>   <char> <num> <num>   <num>  <list> <list>
#> 1: epsilon ParamDbl     0   Inf     Inf       0      0