mlr3mbo: Flexible Bayesian Optimization

A modern and flexible approach to Bayesian Optimization / Model Based Optimization building on the 'bbotk' package. 'mlr3mbo' is a toolbox providing both ready-to-use optimization algorithms as well as their fundamental building blocks allowing for straightforward implementation of custom algorithms. Single- and multi-objective optimization is supported as well as mixed continuous, categorical and conditional search spaces. Moreover, using 'mlr3mbo' for hyperparameter optimization of machine learning models within the 'mlr3' ecosystem is straightforward via 'mlr3tuning'. Examples of ready-to-use optimization algorithms include Efficient Global Optimization by Jones et al. (1998) doi:10.1023/A:1008306431147 , ParEGO by Knowles (2006) doi:10.1109/TEVC.2005.851274 and SMS-EGO by Ponweiser et al. (2008) doi:10.1007/978-3-540-87700-4_78 .

See also

Author

Maintainer: Lennart Schneider lennart.sch@web.de (ORCID)

Authors:

Jakob Richter jakob1richter@gmail.com (ORCID)

Marc Becker marcbecker@posteo.de (ORCID)

Michel Lang michellang@gmail.com (ORCID)

Bernd Bischl bernd_bischl@gmx.net (ORCID)

Florian Pfisterer pfistererf@googlemail.com (ORCID)

Martin Binder mlr.developer@mb706.com

Sebastian Fischer sebf.fischer@gmail.com (ORCID)

Other contributors:

Michael H. Buselli [copyright holder]

Wessel Dankers [copyright holder]

Carlos Fonseca [copyright holder]

Manuel Lopez-Ibanez [copyright holder]

Luis Paquete [copyright holder]