Acquisition Function Expected Improvement Per Second

Expected improvement per second.

It is assumed that calculations are performed on an bbotk::OptimInstanceSingleCrit. Additionally to target values of the codomain that should be minimized or maximized, the bbotk::Objective of the bbotk::OptimInstanceSingleCrit should return time values. The column names of the target variable and time variable must be passed as y_cols in the order (target, time) when constructing the SurrogateLearners that is being used as a surrogate.

Dictionary

This AcqFunction can be instantiated via the dictionary mlr_acqfunctions or with the associated sugar function acqf():

mlr_acqfunctions$get("eips")
acqf("eips")

References

Snoek, Jasper, Larochelle, Hugo, Adams, P R (2012). “Practical Bayesian Optimization of Machine Learning Algorithms.” In Pereira F, Burges CJC, Bottou L, Weinberger KQ (eds.), Advances in Neural Information Processing Systems, volume 25, 2951--2959.

See also

Super classes

bbotk::Objective -> mlr3mbo::AcqFunction -> AcqFunctionEIPS

Public fields

y_best

(numeric(1)).

Active bindings

y_col

(character(1)).

time_col

(character(1)).

Methods

Public methods

AcqFunctionEIPS$new()

AcqFunctionEIPS$update()

AcqFunctionEIPS$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$format()

bbotk::Objective$print()

mlr3mbo::AcqFunction$eval_dt()

mlr3mbo::AcqFunction$eval_many()

Method new()

Creates a new instance of this R6 class.

Usage

AcqFunctionEIPS$new(surrogate = NULL)

Arguments

surrogate

(NULL | SurrogateLearners).

Method update()

Updates acquisition function and sets y_best.

Usage

AcqFunctionEIPS$update()

Method clone()

The objects of this class are cloneable with this method.

Usage

AcqFunctionEIPS$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.