Acquisition Function Probability of Improvement

Probability of Improvement.

Dictionary

This AcqFunction can be instantiated via the dictionary mlr_acqfunctions or with the associated sugar function acqf():

mlr_acqfunctions$get("pi")
acqf("pi")

References

Kushner, J. H (1964). “A New Method of Locating the Maximum Point of an Arbitrary Multipeak Curve in the Presence of Noise.” Journal of Basic Engineering, 86(1), 97--106.

See also

Super classes

bbotk::Objective -> mlr3mbo::AcqFunction -> AcqFunctionPI

Public fields

y_best

(numeric()).

Methods

Public methods

AcqFunctionPI$new()

AcqFunctionPI$update()

AcqFunctionPI$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$format()

bbotk::Objective$print()

mlr3mbo::AcqFunction$eval_dt()

mlr3mbo::AcqFunction$eval_many()

Method new()

Creates a new instance of this R6 class.

Usage

AcqFunctionPI$new(surrogate = NULL)

Arguments

surrogate

(NULL | SurrogateLearner).

Method update()

Updates acquisition function and sets y_best.

Usage

AcqFunctionPI$update()

Method clone()

The objects of this class are cloneable with this method.

Usage

AcqFunctionPI$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.