Sequential Multi-Objective Bayesian Optimization via SMS-EGO

MBO loop function for sequential multi-objective Bayesian Optimization via SmsEGO. Normally used inside an OptimizerMbo.

In each iteration after the initial design, the surrogate and acquisition function (AcqFunctionSmsEgo) are updated and the next candidate is chosen based on optimizing the acquisition function.

Usage

bayesopt_smsego(
  instance,
  init_design_size = NULL,
  surrogate,
  acq_function,
  acq_optimizer,
  random_interleave_iter = 0L
)

Arguments

instance

(bbotk::OptimInstanceMultiCrit)
The bbotk::OptimInstanceMultiCrit to be optimized.

init_design_size

(NULL | integer(1))
Size of the initial design. If NULL and the bbotk::Archive contains no evaluations, 4 * d is used with d being the dimensionality of the search space. Points are drawn uniformly at random.

surrogate

(SurrogateLearnerCollection)
SurrogateLearnerCollection to be used as a surrogate.

acq_function

(AcqFunctionSmsEgo)
AcqFunctionSmsEgo to be used as acquisition function.

acq_optimizer

(AcqOptimizer)
AcqOptimizer to be used as acquisition function optimizer.

random_interleave_iter

(integer(1))
Every random_interleave_iter iteration (starting after the initial design), a point is sampled uniformly at random and evaluated (instead of a model based proposal). For example, if random_interleave_iter = 2, random interleaving is performed in the second, fourth, sixth, ... iteration. Default is 0, i.e., no random interleaving is performed at all.

Value

invisible(instance)

The original instance is modified in-place and returned invisible.

Note

The acq_function$surrogate, even if already populated, will always be overwritten by the surrogate.

The acq_optimizer$acq_function, even if already populated, will always be overwritten by acq_function.

The surrogate$archive, even if already populated, will always be overwritten by the bbotk::Archive of the bbotk::OptimInstanceMultiCrit.

Due to the iterative computation of the epsilon within the AcqFunctionSmsEgo, requires the bbotk::Terminator of the bbotk::OptimInstanceMultiCrit to be a bbotk::TerminatorEvals.

References

Beume N, Naujoks B, Emmerich M (2007). “SMS-EMOA: Multiobjective selection based on dominated hypervolume.” European Journal of Operational Research, 181(3), 1653--1669.

Ponweiser, Wolfgang, Wagner, Tobias, Biermann, Dirk, Vincze, Markus (2008). “Multiobjective Optimization on a Limited Budget of Evaluations Using Model-Assisted S-Metric Selection.” In Proceedings of the 10th International Conference on Parallel Problem Solving from Nature, 784--794.

See also

Examples

if (FALSE) {
if (requireNamespace("mlr3learners") &
    requireNamespace("DiceKriging") &
    requireNamespace("rgenoud")) {

  library(bbotk)
  library(paradox)
  library(mlr3learners)

  fun = function(xs) {
    list(y1 = xs$x^2, y2 = (xs$x - 2) ^ 2)
  }
  domain = ps(x = p_dbl(lower = -10, upper = 10))
  codomain = ps(y1 = p_dbl(tags = "minimize"), y2 = p_dbl(tags = "minimize"))
  objective = ObjectiveRFun$new(fun = fun, domain = domain, codomain = codomain)

  instance = OptimInstanceMultiCrit$new(
    objective = objective,
    terminator = trm("evals", n_evals = 5))

  surrogate = default_surrogate(instance)

  acq_function = acqf("sms_ego")

  acq_optimizer = acqo(
    optimizer = opt("random_search"),
    terminator = trm("evals", n_evals = 100))

  optimizer = opt("mbo",
    loop_function = bayesopt_smsego,
    surrogate = surrogate,
    acq_function = acq_function,
    acq_optimizer = acq_optimizer)

  optimizer$optimize(instance)
}
}