Model Based Optimization

OptimizerMbo class that implements model based optimization. The implementation follows a modular layout relying on a loop function determining the MBO flavor to be used, e.g., bayesopt_ego for sequential single objective MBO, a Surrogate, an AcqFunction, e.g., AcqFunctionEI for expected improvement and an AcqOptimizer.

Super class

bbotk::Optimizer -> OptimizerMbo

Public fields

loop_function

(function | NULL).

surrogate

(Surrogate | NULL)

acq_function

(AcqFunction | NULL).

acq_optimizer

(AcqOptimizer | NULL).

args

(named list() | NULL).

result_function

(function | NULL).

Methods

Public methods

OptimizerMbo$new()

OptimizerMbo$clone()

Inherited methods

bbotk::Optimizer$format()

bbotk::Optimizer$optimize()

bbotk::Optimizer$print()

Method new()

Creates a new instance of this R6 class. For more information on default values for loop_function, surrogate, acq_function and acq_optimizer, see mbo_defaults.

Usage

OptimizerMbo$new(
  loop_function = NULL,
  surrogate = NULL,
  acq_function = NULL,
  acq_optimizer = NULL,
  args = NULL,
  result_function = NULL
)

Arguments

loop_function

(function)
Loop function to run. See mbo_defaults for defaults.

surrogate

Surrogate
The surrogate object. For defaults see mbo_defaults.

acq_function

AcqFunction
The acquistion function object. For defaults see mbo_defaults.

acq_optimizer

AcqOptimizer
The optimizer that optimizes the acquistion function. For defaults see mbo_defaults.

args

(named list())
Further arguments passed to the loop_function.

result_function

(function)
Function called after the optimization terminates. Determines how the final result of the optimization is calculated. Requires arguments inst (the bbotk::OptimInstance) and self (the OptimizerMbo). See for example result_by_surrogate_design which is used by default if the bbotk::OptimInstance has the property "noisy" (which is the case for a mlr3tuning::TuningInstanceSingleCrit or mlr3tuning::TuningInstanceMultiCrit)

Method clone()

The objects of this class are cloneable with this method.

Usage

OptimizerMbo$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.