Acquisition Optimizer

Optimizer for AcqFunction objects.

Public fields

optimizer

(bbotk::Optimizer).

terminator

(bbotk::Terminator).

Active bindings

param_set

(paradox::ParamSet)
Set of hyperparameters.

Methods

Public methods

AcqOptimizer$new()

AcqOptimizer$optimize()

AcqOptimizer$clone()

Method new()

Creates a new instance of this R6 class.

Usage

AcqOptimizer$new(optimizer, terminator)

Arguments

optimizer

(bbotk::Optimizer).

terminator

(bbotk::Terminator).

Method optimize()

Optimize the acquisition function.

If the fix_distance parameter is set to TRUE, proposed points are replaced by randomly sampled ones if their Gower distance with respect to other proposed points or previously evaluated points falls below the dist_threshold parameter. Note that these checks are only performed a single time (and not after a potential replacement of points), i.e., the returned data.table::data.table() of optima must not necessarily pass the checks but most likely will.

Usage

AcqOptimizer$optimize(acq_function, archive)

Arguments

acq_function

AcqFunction
Acquisition function to optimize.

archive

bbotk::Archive
Archive.

Returns

data.table::data.table() with 1 row per optimum and x as columns.

Method clone()

The objects of this class are cloneable with this method.

Usage

AcqOptimizer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.