Surrogate Model Containing Multiple Learners

Surrogate model containing multiple regression mlr3::Learners. The mlr3::Learners are fit on the target variables as indicated via y_cols. Note that redundant mlr3::Learners must be deep clones.

Super class

mlr3mbo::Surrogate -> SurrogateLearnerCollection

Active bindings

print_id

(character)
Id used when printing.

n_learner

(integer(1))
Returns the number of mlr3::Learners.

assert_insample_perf

(numeric())
Asserts whether the current insample performance meets the performance threshold.

packages

(character())
Set of required packages.

feature_types

(character())
Stores the feature types the learner can handle, e.g. "logical", "numeric", or "factor". A complete list of candidate feature types, grouped by task type, is stored in mlr_reflections$task_feature_types.

Methods

Public methods

SurrogateLearnerCollection$new()

SurrogateLearnerCollection$predict()

SurrogateLearnerCollection$clone()

Inherited methods

mlr3mbo::Surrogate$format()

mlr3mbo::Surrogate$print()

mlr3mbo::Surrogate$update()

Method new()

Creates a new instance of this R6 class.

Usage

SurrogateLearnerCollection$new(
  learners,
  archive = NULL,
  x_cols = NULL,
  y_cols = NULL
)

Arguments

learners

(list of mlr3::LearnerRegr).

archive

(NULL | bbotk::Archive).

x_cols

(NULL | character()).

y_cols

(NULL | character()).

Method predict()

Returns a named list of data.tables. Each contains the mean response and standard error for one y_col.

Usage

SurrogateLearnerCollection$predict(xdt)

Arguments

xdt

(data.table::data.table())
New data.

Returns

list of data.table::data.table()s with the columns mean and se.

Method clone()

The objects of this class are cloneable with this method.

Usage

SurrogateLearnerCollection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.