Sequential Singlecriteria Bayesian Optimization Via Multipoint Constant Liar

MBO loop function for sequential singlecriteria Bayesian optimization via multipoint constant liar. Normally used inside an OptimizerMbo.

bayesopt_mpcl(
  instance,
  init_design_size = NULL,
  surrogate = NULL,
  acq_function = NULL,
  acq_optimizer = NULL,
  q = 2L,
  liar = mean,
  random_interleave_iter = 0L
)

Arguments

instance

(bbotk::OptimInstanceSingleCrit)
The bbotk::OptimInstanceSingleCrit to be optimized.

init_design_size

(NULL | integer(1))
Size of the initial design. If NULL 4 * d is used with d being the dimensionality of the search space. Points are drawn uniformly at random.

surrogate

(NULL | Surrogate)
Surrogate to be used as a surrogate. Typically a SurrogateLearner. If NULL default_surrogate(instance) is used.

acq_function

(NULL | AcqFunction). AcqFunction to be used as acquisition function. If NULL default_acqfun(instance) is used.

acq_optimizer

(AcqOptimizer)
AcqOptimizer to be used as acquisition function optimizer. If NULL default_acqopt(acqfun) is used.

q

(integer(1))
Batch size > 1. Default is 2.

liar

(function)
Any function accepting a numeric vector as input and returning a single numeric output. Default is mean.

random_interleave_iter

(integer(1))
Every "random_interleave_iter" iteration (starting after the initial design), a point is sampled uniformly at random and evaluated (instead of a model based proposal). For example, if random_interleave_iter = 2, random interleaving is performed in the second, fourth, sixth, ... iteration. Default is 0, i.e., no random interleaving is performed at all.

Value

invisible(instance)

The original instance is modified in-place and returned invisible.

Note

If surrogate is NULL but acq_function is given and contains a $surrogate, this Surrogate is used.

You can pass a surrogate that was not given the bbotk::Archive of the instance during initialization. In this case, the bbotk::Archive of the given instance is set during execution.

Similarly, you can pass an acq_function that was not given the surrogate during initialization and an acq_optimizer that was not given the acq_function, i.e., delayed initialization is handled automatically.

References

Wang, Jialei, Clark, C. S, Liu, Eric, Frazier, I. P (2020). “Parallel Bayesian Global Optimization of Expensive Functions.” Operations Research, 68(6), 1850--1865.

See also

Examples

library(bbotk)
library(paradox)
library(mlr3learners)

objective = ObjectiveRFun$new(
  fun = function(xs) list(y = xs$x ^ 2),
  domain = ps(x = p_dbl(lower = -5, upper = 5)),
  codomain = ps(y = p_dbl(tags = "minimize"))
)

terminator = trm("evals", n_evals = 5)

instance = OptimInstanceSingleCrit$new(
  objective = objective,
  terminator = terminator
)

bayesopt_mpcl(instance)