Default Surrogate

This is a helper function that generates a default Surrogate based on properties of the bbotk::OptimInstance.

For numeric-only (including integers) parameter spaces without any dependencies:

A Kriging model “"regr.km"” with kernel “"matern3_2"” is created.

If the objective function is deterministic we add a small nugget effect (10^-8*Var(y), y is vector of observed outcomes in current design) to increase numerical stability to hopefully prevent crashes of DiceKriging. Whether the objective function is deterministic can be observed from the objective function's properties.

If the objective function is noisy the nugget effect will be estimated with nugget.estim = TRUE.

Also jitter is set to TRUE to circumvent a problem with DiceKriging where already trained input values produce the exact trained output. Whether the objective function is noisy can be observed from the objective functions properties.

Instead of the default "BFGS" optimization method we use rgenoud ("gen"), which is a hybrid algorithm, to combine global search based on genetic algorithms and local search based on gradients. This may improve the model fit and will less frequently produce a constant surrogate model.

For mixed numeric-categorical parameter spaces, or spaces with conditional parameters:

A ranger regression forest “"regr.ranger"” with 500 trees is created.

The standard error of a prediction (if required by the infill criterion) is estimated via jackknife. This is the se.method = "jack" option of the “"regr.ranger"” learner (default).

In any case, learners are encapsulated using “"evaluate"”, and a fallback learner is set, in cases where the surrogate learner errors. Currently, the following learner is used as a fallback: lrn("regr.ranger", num.trees = 20L, keep.inbag = TRUE, se.method = "jack").

If additionally dependencies are present in the parameter space, inactive conditional parameters are represented by missing NA values in the training design data. We simply handle those with an imputation method, added to the ranger random forest, more concretely we use po("imputesample") (for logicals) and po("imputeoor") (for anything else) from package mlr3pipelines. Characters are always encoded as factors via po("colapply"). Out of range imputation makes sense for tree-based methods and is usually hard to beat, see Ding et al. (2010). In the case of dependencies, the following learner is used as a fallback: lrn("regr.featureless").

If the instance is of class bbotk::OptimInstanceSingleCrit the learner is wrapped as a SurrogateLearner.

If the instance is of class bbotk::OptimInstanceMultiCrit deep clones of the learner are wrapped as a SurrogateLearnerCollection.

Usage

default_surrogate(instance, learner = NULL, n_learner = NULL)

Arguments

instance

(bbotk::OptimInstance)
An object that inherits from bbotk::OptimInstance.

learner

(NULL | mlr3::Learner).

n_learner

(NULL | integer(1)).

Value

Surrogate

References

Ding, Yufeng, Simonoff, S J (2010). “An Investigation of Missing Data Methods for Classification Trees Applied to Binary Response Data.” Journal of Machine Learning Research, 11(1), 131--170.

See also