Acquisition Function SMS EGO

S-Metric Selection Evolutionary Multiobjective Optimization Algorithm.

Parameters

"lambda" (numeric(1))
\(\lambda\) value used for the confidence bound. Defaults to 1. Based on confidence = (1 - 2 * dnorm(lambda)) ^ m you can calculate a lambda for a given confidence level, see Ponweiser et al. 2008.

"epsilon" (numeric(1))
\(\epsilon\) used for the additive epsilon dominance. Can either be a single numeric value > 0 or NULL. In the case of being NULL, an epsilon vector is maintained dynamically as described in Horn et al. 2015.

References

Ponweiser, Wolfgang, Wagner, Tobias, Biermann, Dirk, Vincze, Markus (2008). “Multiobjective Optimization on a Limited Budget of Evaluations Using Model-Assisted S-Metric Selection.” In Proceedings of the 10th International Conference on Parallel Problem Solving from Nature, 784--794. Horn, Daniel, Wagner, Tobias, Biermann, Dirk, Weihs, Claus, Bischl, Bernd (2015). “Model-Based Multi-objective Optimization: Taxonomy, Multi-Point Proposal, Toolbox and Benchmark.” In International Conference on Evolutionary Multi-Criterion Optimization, 64--78.

See also

Super classes

bbotk::Objective -> mlr3mbo::AcqFunction -> AcqFunctionSmsEgo

Public fields

ys_front

(matrix()).

ref_point

(numeric()).

epsilon

(numeric()).

progress

(numeric(1)).

Methods

Public methods

AcqFunctionSmsEgo$new()

AcqFunctionSmsEgo$update()

AcqFunctionSmsEgo$clone()

Inherited methods

bbotk::Objective$eval()

bbotk::Objective$format()

bbotk::Objective$print()

mlr3mbo::AcqFunction$eval_dt()

mlr3mbo::AcqFunction$eval_many()

Method new()

Creates a new instance of this R6 class.

Usage

AcqFunctionSmsEgo$new(surrogate = NULL, lambda = 1, epsilon = NULL)

Arguments

surrogate

(NULL | SurrogateLearners).

lambda

(numeric(1)).

epsilon

(NULL | numeric(1)).

Method update()

Updates acquisition function and sets ys_front, ref_point, epsilon.

Usage

AcqFunctionSmsEgo$update()

Method clone()

The objects of this class are cloneable with this method.

Usage

AcqFunctionSmsEgo$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.