Tuner using Bayesian Optimization

Tune hyperparameters using Bayesian Optimization. This is a minimal interface internally passing on to OptimizerMbo. For additional information and documentation see OptimizerMbo.

Super classes

mlr3tuning::Tuner -> mlr3tuning::TunerFromOptimizer -> TunerMbo

Active bindings

loop_function

(function | NULL).

surrogate

(Surrogate | NULL)

acq_function

(AcqFunction | NULL).

acq_optimizer

(AcqOptimizer | NULL).

args

(named list() | NULL).

result_function

(function | NULL).

Methods

Public methods

TunerMbo$new()

TunerMbo$clone()

Inherited methods

mlr3tuning::Tuner$format()

mlr3tuning::Tuner$help()

mlr3tuning::Tuner$print()

mlr3tuning::TunerFromOptimizer$optimize()

Method new()

Creates a new instance of this R6 class. For more information on default values for loop_function, surrogate, acq_function and acq_optimizer, see mbo_defaults.

Note that all the parameters below are simply passed to the OptimizerMbo and the respective fields are simply (settable) active bindings to the fields of the OptimizerMbo.

Usage

TunerMbo$new(
  loop_function = NULL,
  surrogate = NULL,
  acq_function = NULL,
  acq_optimizer = NULL,
  args = NULL,
  result_function = NULL
)

Arguments

loop_function

(function)
Loop function to run. See mbo_defaults for defaults.

surrogate

Surrogate
The surrogate object. For defaults see mbo_defaults.

acq_function

AcqFunction
The acquistion function object. For defaults see mbo_defaults.

acq_optimizer

AcqOptimizer
The optimizer that optimizes the acquistion function. For defaults see mbo_defaults.

args

(named list())
Further arguments passed to the loop_function. For example, random_interleave_iter.

result_function

(function)
Optional function called after the optimization terminates. Determines how the final result of the optimization is calculated. Requires arguments inst (the bbotk::OptimInstance) and self (the OptimizerMbo). See for example result_by_surrogate_design which is used by default if the bbotk::OptimInstance has the property "noisy" (which is the case for a mlr3tuning::TuningInstanceSingleCrit or mlr3tuning::TuningInstanceMultiCrit).

Method clone()

The objects of this class are cloneable with this method.

Usage

TunerMbo$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.